Update unique_id in login API to include app name and version
The code change updates the `unique_id` field in the request body for the login API call. Previously, the `unique_id` was set to `_userAgent`. Now, it is set to a string that combines the name and version of the entry assembly, formatted as "AppName vVersion".

This modification ensures that the `unique_id` sent in the login request is more descriptive, including both the application name and its version, which can be useful for tracking and debugging purposes.